# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PI Planning API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/pis/{pi_id}/swimlines": {
      "get": {
        "tags": [
          "swimlines"
        ],
        "summary": "List Swimlines",
        "operationId": "list_swimlines_api_v1_pis__pi_id__swimlines_get",
        "parameters": [
          {
            "name": "pi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pi Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SwimlineResponse"
                  },
                  "title": "Response List Swimlines Api V1 Pis  Pi Id  Swimlines Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "swimlines"
        ],
        "summary": "Create Swimline",
        "operationId": "create_swimline_api_v1_pis__pi_id__swimlines_post",
        "parameters": [
          {
            "name": "pi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pi Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwimlineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwimlineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swimlines/{swimline_id}": {
      "get": {
        "tags": [
          "swimlines"
        ],
        "summary": "Get Swimline",
        "operationId": "get_swimline_api_v1_swimlines__swimline_id__get",
        "parameters": [
          {
            "name": "swimline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swimline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwimlineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "swimlines"
        ],
        "summary": "Update Swimline",
        "operationId": "update_swimline_api_v1_swimlines__swimline_id__patch",
        "parameters": [
          {
            "name": "swimline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swimline Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwimlineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwimlineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "swimlines"
        ],
        "summary": "Delete Swimline",
        "operationId": "delete_swimline_api_v1_swimlines__swimline_id__delete",
        "parameters": [
          {
            "name": "swimline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Swimline Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pis/{pi_id}/sprints": {
      "get": {
        "tags": [
          "sprints"
        ],
        "summary": "List Sprints",
        "operationId": "list_sprints_api_v1_pis__pi_id__sprints_get",
        "parameters": [
          {
            "name": "pi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pi Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SprintResponse"
                  },
                  "title": "Response List Sprints Api V1 Pis  Pi Id  Sprints Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sprints"
        ],
        "summary": "Create Sprint",
        "operationId": "create_sprint_api_v1_pis__pi_id__sprints_post",
        "parameters": [
          {
            "name": "pi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pi Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SprintCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sprints/{sprint_id}": {
      "patch": {
        "tags": [
          "sprints"
        ],
        "summary": "Update Sprint",
        "operationId": "update_sprint_api_v1_sprints__sprint_id__patch",
        "parameters": [
          {
            "name": "sprint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sprint Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SprintUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SprintCreate": {
        "properties": {
          "sprint_index": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Sprint Index"
          },
          "capacity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Capacity"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "sprint_index",
          "capacity"
        ],
        "title": "SprintCreate"
      },
      "SprintResponse": {
        "properties": {
          "system_id": {
            "type": "string",
            "title": "System Id"
          },
          "pi_id": {
            "type": "string",
            "title": "Pi Id"
          },
          "sprint_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprint Index"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "modified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Modified At"
          }
        },
        "type": "object",
        "required": [
          "system_id",
          "pi_id",
          "sprint_index",
          "capacity",
          "start_date",
          "end_date",
          "created_at",
          "modified_at"
        ],
        "title": "SprintResponse"
      },
      "SprintUpdate": {
        "properties": {
          "capacity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "title": "SprintUpdate"
      },
      "SwimlineCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "order_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Index"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SwimlineCreate"
      },
      "SwimlineResponse": {
        "properties": {
          "system_id": {
            "type": "string",
            "title": "System Id"
          },
          "pi_id": {
            "type": "string",
            "title": "Pi Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "order_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Index"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "modified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Modified At"
          }
        },
        "type": "object",
        "required": [
          "system_id",
          "pi_id",
          "name",
          "order_index",
          "created_at",
          "modified_at"
        ],
        "title": "SwimlineResponse"
      },
      "SwimlineUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "order_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Index"
          }
        },
        "type": "object",
        "title": "SwimlineUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}